refactor: group configs in config folder
