+++ src/lib/config/refresh-intervals.ts
