+++ src/lib/utils.ts
@@ -1,6 +1,6 @@
 import { clsx, type ClassValue } from 'clsx'
 import { twMerge } from 'tailwind-merge'
-import { QUBIC_EXPLORER_BASE_URL } from './constants'
+import { QUBIC_EXPLORER_BASE_URL } from './config/constants'
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs))
