+++ src/pages/history.tsx
@@ -19,7 +19,7 @@ import {
 } from '@/lib/pending-transactions'
 import { getCurrentIdentity } from '@/lib/accounts'
 import HistoryEmptyState from '@/components/pages/history/history-empty-state'
-import { REFRESH_INTERVAL_ACTIVE_TRANSACTIONS } from '@/lib/refresh-intervals'
+import { REFRESH_INTERVAL_ACTIVE_TRANSACTIONS } from '@/lib/config/refresh-intervals'
 
 const HistoryRowSkeleton = () => (
   <div className="space-y-3 rounded-xl border border-border/40 bg-background/40 px-3 py-3">
